+++ CHANGELOG.md
@@ -83,6 +83,7 @@ Observing messages from data track
 Improvements
 
 - Moved pre-defined aspect ratios from `VideoConstraints` class to `AspectRatio` class.
+- Local audio, video, and data tracks return their track IDs for `getName` if no name was specified.
     
 Bug Fixes